splited the index.ts file and put the app in a seperate app.ts file for further uses like creating testing environment..
