+++ auth/Dockerfile
@@ -4,7 +4,7 @@ WORKDIR /app
 
 COPY ./package.json ./
 
-RUN npm install
+RUN npm install --omit=dev
 
 COPY ./ ./
