+++ auth/src/index.ts
@@ -1,35 +1,6 @@
-import express from 'express';
-import 'express-async-errors';
 import mongoose from 'mongoose';
-import cookieSession from 'cookie-session';
 
-import { currentUserRouter } from './routes/current-user';
-import { signupRouter } from './routes/signup';
-import { signinRouter } from './routes/signin';
-import { signoutRouter } from './routes/signout';
-import { errorHandler } from './middlewares/error-handler';
-import { NotFoundError } from './errors/not-found-error';
-
-const app = express();
-app.use(express.json());
-app.set('trust proxy', true);
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true,
-  })
-);
-
-app.use(currentUserRouter);
-app.use(signupRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-
-app.all('*', () => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
+import { app } from './app'
 
 const start = async () => {
   if (!process.env.JWT_KEY) {
